+++ app/src/main/java/com/arielwang/workoutlogger/database/AppDatabase.kt
@@ -1,25 +1,26 @@
 package com.arielwang.workoutlogger.database
 
-import androidx.room.AutoMigration
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import androidx.room.TypeConverters
+import androidx.room.*
+import androidx.room.migration.AutoMigrationSpec
 import com.arielwang.workoutlogger.database.converters.ExerciseTypeConverter
 import com.arielwang.workoutlogger.database.daos.ExerciseCardDao
-import com.arielwang.workoutlogger.database.daos.ExerciseDao
+import com.arielwang.workoutlogger.database.daos.WorkoutDao
 import com.arielwang.workoutlogger.database.model.ExerciseCard
-import com.arielwang.workoutlogger.database.model.ExerciseFlow
+import com.arielwang.workoutlogger.database.model.WorkoutAddingFlow
 
-const val DATABASE_VERSION = 3
+const val DATABASE_VERSION = 4
 @Database(
-  entities = [ExerciseFlow::class, ExerciseCard::class],
+  entities = [WorkoutAddingFlow::class, ExerciseCard::class],
   version = DATABASE_VERSION,
   autoMigrations = [
-    AutoMigration (from = 2, to = 3)
+    AutoMigration (from = 3, to = 4, spec = AppDatabase.MyAutoMigration::class)
   ]
 )
 @TypeConverters(ExerciseTypeConverter::class)
 abstract class AppDatabase : RoomDatabase() {
-  abstract fun exerciseDao(): ExerciseDao
+  @RenameTable(fromTableName = "ExerciseFlow", toTableName = "WorkoutAddingFlow")
+  class MyAutoMigration : AutoMigrationSpec
+
+  abstract fun exerciseDao(): WorkoutDao
   abstract fun exerciseCardDao(): ExerciseCardDao
 }
