--- app/src/main/java/com/arielwang/workoutlogger/database/daos/ExerciseDao.kt
@@ -1,15 +1,0 @@
-package com.arielwang.workoutlogger.database.daos
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-import com.arielwang.workoutlogger.database.model.ExerciseFlow
-
-@Dao
-interface ExerciseDao {
-  @Query("SELECT * FROM ExerciseFlow")
-  suspend fun getAllExercise(): List<ExerciseFlow>
-
-  @Insert
-  suspend fun insertExercise(exercise: ExerciseFlow)
-}
