--- app/src/main/java/com/arielwang/workoutlogger/features/addexerciseflow/shared/domain/ExerciseSharedStateManager.kt
@@ -1,12 +1,0 @@
-package com.arielwang.workoutlogger.features.addexerciseflow.shared.domain
-
-import com.arielwang.workoutlogger.database.model.ExerciseFlow
-
-interface ExerciseSharedStateManager {
-
-    fun updateState(state: ExerciseFlow)
-
-    fun getState(): ExerciseFlow
-
-    fun deleteState()
-}
