--- app/src/main/java/com/arielwang/workoutlogger/features/addexerciseflow/track/data/di/TrackModule.kt
@@ -1,17 +1,0 @@
-package com.arielwang.workoutlogger.features.addexerciseflow.track.data.di
-
-import com.arielwang.workoutlogger.features.addexerciseflow.track.data.repository.TrackRepositoryImpl
-import com.arielwang.workoutlogger.features.addexerciseflow.track.domain.repository.TrackRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@InstallIn(SingletonComponent::class)
-@Module
-internal abstract class TrackModule {
-    @Binds
-    abstract fun bindTrackRepository(
-        impl: TrackRepositoryImpl,
-    ): TrackRepository
-}
