--- app/src/main/java/com/arielwang/workoutlogger/features/addexerciseflow/track/domain/repository/TrackRepository.kt
@@ -1,7 +1,0 @@
-package com.arielwang.workoutlogger.features.addexerciseflow.track.domain.repository
-
-import com.arielwang.workoutlogger.database.model.ExerciseFlow
-
-interface TrackRepository {
-    suspend fun insertExercise(exercise: ExerciseFlow)
-}
