+++ app/src/main/java/com/arielwang/workoutlogger/features/home/domain/repository/HomeRepository.kt
@@ -1,8 +1,8 @@
 package com.arielwang.workoutlogger.features.home.domain.repository
 
-import com.arielwang.workoutlogger.database.model.ExerciseFlow
+import com.arielwang.workoutlogger.database.model.WorkoutAddingFlow
 
 
 interface HomeRepository {
-  suspend fun getAllExercises(): List<ExerciseFlow>
+  suspend fun getAllExercises(): List<WorkoutAddingFlow>
 }
