+++ app/src/main/java/com/arielwang/workoutlogger/features/workoutaddingflow/exercise/domain/ExerciseRepository.kt
@@ -1,0 +1,7 @@
+package com.arielwang.workoutlogger.features.workoutaddingflow.exercise.domain
+
+import com.arielwang.workoutlogger.database.model.ExerciseCard
+
+interface ExerciseRepository {
+    suspend fun getAllExerciseCards(): List<ExerciseCard>
+}
