+++ app/src/main/java/com/arielwang/workoutlogger/features/workoutaddingflow/track/domain/repository/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.arielwang.workoutlogger.features.workoutaddingflow.track.domain.repository
+
+import com.arielwang.workoutlogger.database.model.WorkoutAddingFlow
+
+interface TrackRepository {
+    suspend fun insertExercise(exercise: WorkoutAddingFlow)
+}
