+++ app/src/main/java/com/arielwang/workoutlogger/root/RootScreen.kt
@@ -4,11 +4,12 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
-import com.arielwang.workoutlogger.features.addexerciseflow.exercise.ui.screen.addExerciseScreen
+import com.arielwang.workoutlogger.features.workoutaddingflow.exercise.ui.screen.addExerciseScreen
 import com.arielwang.workoutlogger.features.home.ui.screen.HomeDestination
 import com.arielwang.workoutlogger.features.home.ui.screen.addHomeScreen
 import com.arielwang.workoutlogger.features.landing.ui.screen.addLandingScreen
-import com.arielwang.workoutlogger.features.addexerciseflow.track.ui.screen.addTrackScreen
+import com.arielwang.workoutlogger.features.workoutaddingflow.track.ui.screen.addTrackScreen
+import com.arielwang.workoutlogger.features.exercisecard.ui.addExerciseDetailScreen
 import com.google.accompanist.navigation.material.BottomSheetNavigator
 import com.google.accompanist.navigation.material.ExperimentalMaterialNavigationApi
 
@@ -27,6 +28,7 @@ fun RootScreen(
     addHomeScreen()
     addLandingScreen()
     addExerciseScreen()
+    addExerciseDetailScreen()
     addTrackScreen()
   }
 }
