+++ app/src/main/res/values/strings.xml
@@ -5,14 +5,15 @@
     <string name="LandingScreen_buttonContent">Next Screen</string>
     <string name="HomeScreen_title">No Workout For Today</string>
     <string name="HomeScreen_buttonContent">Create New Workout</string>
-    <string name="ExerciseCard_abs">abs</string>
-    <string name="ExerciseCard_back">Back</string>
-    <string name="ExerciseCard_biceps">Biceps</string>
-    <string name="ExerciseCard_cardio">Cardio</string>
-    <string name="ExerciseCard_chest">Chest</string>
     <string name="ExerciseScreen_buttonContent">Next</string>
-    <string name="ExerciseScreen_topbarIconButtonContentDescription">BackIcon</string>
+    <string name="ExerciseScreen_topbarBackIconButton">BackIcon</string>
+    <string name="ExerciseScreen_topbarAddIconButton">AddIcon</string>
     <string name="ExerciseScreen_topbarTitle">Exercise</string>
+    <string name="ExerciseDetailScreen_topbarTitle">New Exercise</string>
+    <string name="ExerciseDetailScreen_topbarBackIconButton">BackIcon</string>
+    <string name="ExerciseDetailScreen_topbarSaveIconButton">SaveIcon</string>
+    <string name="ExerciseDetailScreen_nameSection">Name</string>
+    <string name="ExerciseDetailScreen_notesSection">Notes</string>
     <string name="TrackScreen_topbarIconButtonContentDescription">BackIcon</string>
     <string name="TrackScreen_topbarTitle">Track</string>
     <string name="TrackScreen_buttonContent">Save</string>
